Added Github Update check
