+++ public/assets/XXL-WebApp-d09ec568f7e96b796d488cf060517436.js
@@ -1,0 +1,63 @@
+"use strict"
+define("XXL-WebApp/app",["exports","XXL-WebApp/resolver","ember-load-initializers","XXL-WebApp/config/environment"],function(e,t,n,a){Object.defineProperty(e,"__esModule",{value:!0})
+var s=Ember.Application.extend({modulePrefix:a.default.modulePrefix,podModulePrefix:a.default.podModulePrefix,rootElement:".app",environment:a.default.environment,Resolver:t.default});(0,n.default)(s,a.default.modulePrefix),e.default=s}),define("XXL-WebApp/components/data-display",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Component.extend({tagName:"dl",classNames:["data-display"]})}),define("XXL-WebApp/components/nav-panel",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Component.extend({})}),define("XXL-WebApp/components/odn-network",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Component.extend({})}),define("XXL-WebApp/components/odn-node",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Component.extend({})}),define("XXL-WebApp/components/odn-server",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Component.extend({})}),define("XXL-WebApp/components/side-bar",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Component.extend({})}),define("XXL-WebApp/components/token-value-data-list",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Component.extend({tagName:"span"})}),define("XXL-WebApp/components/xxl-footer",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Component.extend({tagName:"footer",classNames:["footer","monospace"]})}),define("XXL-WebApp/components/xxl-header",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Component.extend({tagName:"header",classNames:["header"],actions:{toggleMobileNav:function(){this.sendAction("toggleMobileNav")},search:function(e){this.sendAction("search",e)}}})}),define("XXL-WebApp/controllers/application",["exports","XXL-WebApp/config/environment"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Controller.extend({odnNetwork:Ember.inject.service(),xxlVersion:t.default.version,isSearching:!1,isMobileNavActive:!1,actions:{search:function(e){var t=this
+!1===this.get("isSearching")&&(this.toggleProperty("isSearching"),this.get("odnNetwork").runSearch(e).then(function(n){"success"==n.status?("block"===n.response.search_type?t.transitionToRoute("block",e):"transaction"===n.response.search_type?t.transitionToRoute("transaction",e):"address"===n.response.search_type?t.transitionToRoute("address",e):t.transitionToRoute("search",{queryParams:{query:e}}),Ember.run.later(function(){console.log("finished search"),t.toggleProperty("isSearching")},2e3)):"error"==n.status&&(t.transitionToRoute("search",{queryParams:{query:e}}),Ember.run.later(function(){t.toggleProperty("isSearching")},2e3))}).catch(function(e){console.warn("Search error occurred",e),Ember.run.later(function(){t.toggleProperty("isSearching")},2e3)}))},toggleMobileNav:function(e){this.toggleProperty("isMobileNavActive"),!0===this.get("isMobileNavActive")?$("body").addClass("nav-open"):$("body").removeClass("nav-open")}}})}),define("XXL-WebApp/helpers/and",["exports","ember-truth-helpers/helpers/and"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"and",{enumerable:!0,get:function(){return t.and}})}),define("XXL-WebApp/helpers/app-version",["exports","XXL-WebApp/config/environment","ember-cli-app-version/utils/regexp"],function(e,t,n){function a(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{}
+return t.hideSha?s.match(n.versionRegExp)[0]:t.hideVersion?s.match(n.shaRegExp)[0]:s}Object.defineProperty(e,"__esModule",{value:!0}),e.appVersion=a
+var s=t.default.APP.version
+e.default=Ember.Helper.helper(a)}),define("XXL-WebApp/helpers/display-time-diff",["exports"],function(e){function t(e){return e}Object.defineProperty(e,"__esModule",{value:!0}),e.displayTimeDiff=t,e.default=Ember.Helper.helper(t)}),define("XXL-WebApp/helpers/eq",["exports","ember-truth-helpers/helpers/equal"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"equal",{enumerable:!0,get:function(){return t.equal}})}),define("XXL-WebApp/helpers/format-bytes",["exports"],function(e){function t(e,t){var a=n(e,1),s=a[0]
+s=parseInt(s)
+var r=-1,l=["kB","MB","GB","TB","PB","EB","ZB","YB"]
+do{s/=1024,r++}while(s>1024)
+var o=Math.max(s,.1).toFixed(1),i=l[r]
+return t.unitWrap?Ember.String.htmlSafe(o+' <span class="shrink muted">'+i+"</span>"):o+" "+i}Object.defineProperty(e,"__esModule",{value:!0}),e.formatBytes=t
+var n=function(){function e(e,t){var n=[],a=!0,s=!1,r=void 0
+try{for(var l,o=e[Symbol.iterator]();!(a=(l=o.next()).done)&&(n.push(l.value),!t||n.length!==t);a=!0);}catch(e){s=!0,r=e}finally{try{!a&&o.return&&o.return()}finally{if(s)throw r}}return n}return function(t,n){if(Array.isArray(t))return t
+if(Symbol.iterator in Object(t))return e(t,n)
+throw new TypeError("Invalid attempt to destructure non-iterable instance")}}()
+e.default=Ember.Helper.helper(t)}),define("XXL-WebApp/helpers/format-currency",["exports"],function(e){function t(e,t){var n=void 0===t.decimals?2:t.decimals
+return Math.abs(e).toFixed(n)}Object.defineProperty(e,"__esModule",{value:!0}),e.formatCurrency=t,e.default=Ember.Helper.helper(t)}),define("XXL-WebApp/helpers/format-node-version",["exports"],function(e){function t(e){return"string"!=typeof e&&(e=e[0]),e.split("-")[0]}Object.defineProperty(e,"__esModule",{value:!0}),e.formatNodeVersion=t,e.default=Ember.Helper.helper(t)}),define("XXL-WebApp/helpers/format-staking-time",["exports"],function(e){function t(e){var t=parseInt(e)
+return isNaN(t)?e:moment.duration(t,"seconds").humanize(!0)}Object.defineProperty(e,"__esModule",{value:!0}),e.formatStakingTime=t,e.default=Ember.Helper.helper(t)}),define("XXL-WebApp/helpers/format-staking-weight",["exports"],function(e){function t(e){if(isNaN(e))return e
+var t=parseFloat(e/1e8)
+return t=t.toFixed(2),t>1e6&&(t=t.toString().replace(/\B(?=(\d{3})+(?!\d))/g,","),t+=' <span class="muted shrink">M</span>'),Ember.String.htmlSafe(t)}Object.defineProperty(e,"__esModule",{value:!0}),e.formatStakingWeight=t,e.default=Ember.Helper.helper(t)}),define("XXL-WebApp/helpers/format-timestamp",["exports"],function(e){function t(e){var t=n(e,1),s=t[0]
+if(isNaN(s))return"NaN"
+s*=1e3
+var r=new Date(s),l=r.getMinutes()
+l<10&&(l="0"+l)
+var o=r.getHours()>12?"pm":"am",i=r.getHours()
+i>12&&(i-=12)
+var d='<span class="shrink muted xx-sm-block">'+i+":"+l+" "+o,c=a[r.getMonth()]+" "+r.getDate()+", "+r.getFullYear()
+return Ember.String.htmlSafe(c+" "+d)}Object.defineProperty(e,"__esModule",{value:!0}),e.formatTimestamp=t
+var n=function(){function e(e,t){var n=[],a=!0,s=!1,r=void 0
+try{for(var l,o=e[Symbol.iterator]();!(a=(l=o.next()).done)&&(n.push(l.value),!t||n.length!==t);a=!0);}catch(e){s=!0,r=e}finally{try{!a&&o.return&&o.return()}finally{if(s)throw r}}return n}return function(t,n){if(Array.isArray(t))return t
+if(Symbol.iterator in Object(t))return e(t,n)
+throw new TypeError("Invalid attempt to destructure non-iterable instance")}}(),a=["Jan","Feb","Mar","Apr","May","Jun","Jul","Aug","Sept","Oct","Nov","Dec"]
+e.default=Ember.Helper.helper(t)}),define("XXL-WebApp/helpers/format-unix",["exports"],function(e){function t(e){var t=new Date(1e3*e)
+return t.getMonth()+1+"/"+t.getDate()+"/"+t.getFullYear()+" "+t.getHours()+":"+t.getMinutes()}Object.defineProperty(e,"__esModule",{value:!0}),e.formatUnix=t,e.default=Ember.Helper.helper(t)}),define("XXL-WebApp/helpers/gt",["exports","ember-truth-helpers/helpers/gt"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"gt",{enumerable:!0,get:function(){return t.gt}})}),define("XXL-WebApp/helpers/gte",["exports","ember-truth-helpers/helpers/gte"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"gte",{enumerable:!0,get:function(){return t.gte}})}),define("XXL-WebApp/helpers/increment-index",["exports"],function(e){function t(e){return parseInt(e)+1}Object.defineProperty(e,"__esModule",{value:!0}),e.incrementIndex=t,e.default=Ember.Helper.helper(t)}),define("XXL-WebApp/helpers/is-array",["exports","ember-truth-helpers/helpers/is-array"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"isArray",{enumerable:!0,get:function(){return t.isArray}})}),define("XXL-WebApp/helpers/is-equal",["exports","ember-truth-helpers/helpers/is-equal"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"isEqual",{enumerable:!0,get:function(){return t.isEqual}})}),define("XXL-WebApp/helpers/lt",["exports","ember-truth-helpers/helpers/lt"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"lt",{enumerable:!0,get:function(){return t.lt}})}),define("XXL-WebApp/helpers/lte",["exports","ember-truth-helpers/helpers/lte"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"lte",{enumerable:!0,get:function(){return t.lte}})}),define("XXL-WebApp/helpers/not-eq",["exports","ember-truth-helpers/helpers/not-equal"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"notEq",{enumerable:!0,get:function(){return t.notEq}})})
+define("XXL-WebApp/helpers/not",["exports","ember-truth-helpers/helpers/not"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"not",{enumerable:!0,get:function(){return t.not}})}),define("XXL-WebApp/helpers/or",["exports","ember-truth-helpers/helpers/or"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"or",{enumerable:!0,get:function(){return t.or}})}),define("XXL-WebApp/helpers/pluralize",["exports","ember-inflector/lib/helpers/pluralize"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=t.default}),define("XXL-WebApp/helpers/singularize",["exports","ember-inflector/lib/helpers/singularize"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=t.default}),define("XXL-WebApp/helpers/xor",["exports","ember-truth-helpers/helpers/xor"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"xor",{enumerable:!0,get:function(){return t.xor}})}),define("XXL-WebApp/initializers/allow-link-action",["exports","ember-link-action/initializers/allow-link-action"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}}),Object.defineProperty(e,"initialize",{enumerable:!0,get:function(){return t.initialize}})}),define("XXL-WebApp/initializers/app-version",["exports","ember-cli-app-version/initializer-factory","XXL-WebApp/config/environment"],function(e,t,n){Object.defineProperty(e,"__esModule",{value:!0})
+var a=n.default.APP,s=a.name,r=a.version
+e.default={name:"App Version",initialize:(0,t.default)(s,r)}}),define("XXL-WebApp/initializers/container-debug-adapter",["exports","ember-resolver/resolvers/classic/container-debug-adapter"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"container-debug-adapter",initialize:function(){var e=arguments[1]||arguments[0]
+e.register("container-debug-adapter:main",t.default),e.inject("container-debug-adapter:main","namespace","application:main")}}}),define("XXL-WebApp/initializers/data-adapter",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"data-adapter",before:"store",initialize:function(){}}}),define("XXL-WebApp/initializers/ember-data",["exports","ember-data/setup-container","ember-data"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"ember-data",initialize:t.default}}),define("XXL-WebApp/initializers/export-application-global",["exports","XXL-WebApp/config/environment"],function(e,t){function n(){var e=arguments[1]||arguments[0]
+if(!1!==t.default.exportApplicationGlobal){var n
+if("undefined"!=typeof window)n=window
+else if("undefined"!=typeof global)n=global
+else{if("undefined"==typeof self)return
+n=self}var a,s=t.default.exportApplicationGlobal
+a="string"==typeof s?s:Ember.String.classify(t.default.modulePrefix),n[a]||(n[a]=e,e.reopen({willDestroy:function(){this._super.apply(this,arguments),delete n[a]}}))}}Object.defineProperty(e,"__esModule",{value:!0}),e.initialize=n,e.default={name:"export-application-global",initialize:n}}),define("XXL-WebApp/initializers/injectStore",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"injectStore",before:"store",initialize:function(){}}}),define("XXL-WebApp/initializers/store",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"store",after:"ember-data",initialize:function(){}}}),define("XXL-WebApp/initializers/transforms",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"transforms",before:"store",initialize:function(){}}}),define("XXL-WebApp/instance-initializers/ember-data",["exports","ember-data/instance-initializers/initialize-store-service"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default={name:"ember-data",initialize:t.default}}),define("XXL-WebApp/mixins/link-action",["exports","ember-link-action/mixins/link-action"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("XXL-WebApp/models/block",["exports","ember-data"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=t.default.Model.extend({hash:t.default.attr("string"),time:t.default.attr("number"),mint:t.default.attr("number"),size:t.default.attr("number"),flags:t.default.attr("string"),height:t.default.attr("number"),difficulty:t.default.attr("number"),signature:t.default.attr("string"),merkleroot:t.default.attr("string"),previousblockhash:t.default.attr("string"),nextblockhash:t.default.attr("string"),tx:t.default.attr()})}),define("XXL-WebApp/models/peer",["exports","ember-data"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=t.default.Model.extend({addr:t.default.attr(),services:t.default.attr(),lastsend:t.default.attr(),lastrecv:t.default.attr(),bytessent:t.default.attr(),bytesrecv:t.default.attr(),conntime:t.default.attr(),timeoffset:t.default.attr(),pingtime:t.default.attr(),version:t.default.attr(),subver:t.default.attr(),inbound:t.default.attr("boolean"),startingheight:t.default.attr(),banscore:t.default.attr(),syncnode:t.default.attr(),nodever:Ember.computed("subver",function(){return this.get("subver").split(":")[1]}),nodeip:Ember.computed("addr",function(){return this.get("addr").split(":")[0]})})}),define("XXL-WebApp/resolver",["exports","ember-resolver"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=t.default}),define("XXL-WebApp/router",["exports","XXL-WebApp/config/environment"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0})
+var n=Ember.Router.extend({location:t.default.locationType,rootURL:t.default.rootURL})
+n.map(function(){t.default.APP.enable_search&&(this.route("block",{path:"/block/:block_id"}),this.route("transaction",{path:"/transaction/:txid"}),this.route("address",{path:"/address/:address_key"}),this.route("search",{queryParams:["query"]})),this.route("about"),this.route("stats"),this.route("history"),this.route("error",{path:"*:"})}),e.default=n}),define("XXL-WebApp/routes/about",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Route.extend({})}),define("XXL-WebApp/routes/address",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0})
+var t=function(e,t){return new Ember.RSVP.Promise(function(n,a){t.fetchAddress(e).then(function(e){"success"==e.status?n(e.response):a("error"==e.status?e.error:"An error occurred communicating with the network.")}).catch(a)})}
+e.default=Ember.Route.extend({odnNetwork:Ember.inject.service(),model:function(e){return Ember.RSVP.hash({params:e,name:"sample",response:t(e.address_key,this.get("odnNetwork"))}).catch(function(e){return console.error("We ran into an issue... monkeys will be dispatched to your location.",e),{msg:"Recovered from rejected promise",error:e}})},actions:{error:function(e){console.error("We ran into an issue... monkeys will be dispatched to your location.",e),this.transitionTo("index")}}})}),define("XXL-WebApp/routes/application",["exports","XXL-WebApp/config/environment"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0})
+var n=function(){return new Promise(function(e,n){Ember.$.getJSON("https://api.github.com/repos/manbearpixel/ObsidianXXL/releases/latest").then(function(n){var a=t.default.version,s=n.tag_name.replace(/[^0-9.]/g,""),r=!1
+a<s&&(r=!0),e({update:r,version:s,link:n.html_url})}).catch(function(t){e({update:!1})})})}
+e.default=Ember.Route.extend({odnNetwork:Ember.inject.service(),beforeModel:function(){!1===t.default.APP.enable_search&&$("body").addClass("no-search")},model:function(){return Ember.RSVP.hash({priceInfo:this.get("odnNetwork").getPriceInfo(),odnDataInfo:this.get("odnNetwork").getInfo(),odnDataMining:this.get("odnNetwork").getMiningInfo(),currentBlock:"--",networkWeight:"--",enableSearch:t.default.APP.enable_search,latestRelease:n()})},actions:{didTransition:function(){},loading:function(e,t){return!0},error:function(e,t){0===e.status?console.warn("Sorry, but we're having trouble connecting to the server. This problem is usually the result of a broken Internet connection. You can try refreshing this page."):403==e.status?console.warn("error 403"):401==e.status?console.warn("error 401"):404==e.status?this.transitionTo("error"):console.warn("ERROR",e)}},afterModel:function(e){}})}),define("XXL-WebApp/routes/block",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0})
+var t=function(e,t){return new Ember.RSVP.Promise(function(n,a){t.fetchBlock(e).then(function(e){console.log("response get!",e.response),"success"==e.status?n(e.response):a("error"==e.status?e.error:"An error occurred communicating with the network.")}).catch(a)})}
+e.default=Ember.Route.extend({odnNetwork:Ember.inject.service(),model:function(e){return console.log("search"),Ember.RSVP.hash({params:e,name:"sample",passedBlock:{},response:t(e.block_id,this.get("odnNetwork"))}).catch(function(e){return console.error("We ran into an issue... monkeys will be dispatched to your location.",e),{msg:"Recovered from rejected promise",error:e}})},title:"foobar",actions:{error:function(e){console.error("We ran into an issue... monkeys will be dispatched to your location.",e)}}})}),define("XXL-WebApp/routes/error",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Route.extend({})}),define("XXL-WebApp/routes/history",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Route.extend({})}),define("XXL-WebApp/routes/index",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Route.extend({odnNetwork:Ember.inject.service(),model:function(){return Ember.RSVP.hash({systemInfo:this.get("odnNetwork").getSystemInfo(),odnDataInfo:this.get("odnNetwork").getInfo(),odnDataStaking:this.get("odnNetwork").getStakingInfo(),odnDataNetwork:this.get("odnNetwork").getNetworkInfo()})},actions:{loading:function(e,t){return!0},error:function(e,t){return console.warn("INDEX ERROR",e),!0}},events:{error:function(e,t){console.warn("Index Error!"),console.warn("Reason",e)}}})}),define("XXL-WebApp/routes/search",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Route.extend({odnNetwork:Ember.inject.service(),queryParams:{query:""},beforeModel:function(e){this._super.apply(this,arguments),console.log("beforeModel")},setupController:function(e,t){this._super(e,t),console.log("setupController")},model:function(e){return console.log("params",e),Ember.RSVP.hash({name:"sample",params:e}).catch(function(e){return console.error("We ran into an issue... monkeys will be dispatched to your location.",e),{msg:"Recovered from rejected promise",error:e}})},actions:{error:function(e){console.error("We ran into an issue... monkeys will be dispatched to your location.",e)}}})}),define("XXL-WebApp/routes/stats",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.Route.extend({})}),define("XXL-WebApp/routes/test",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0})
+Ember.inject
+e.default=Ember.Route.extend({})})
+define("XXL-WebApp/routes/transaction",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0})
+var t=function(e,t){return new Ember.RSVP.Promise(function(n,a){console.log("fetchTransaction",e),t.fetchTransaction(e).then(function(e){console.log("response get!",e),"success"==e.status?n(e.response):a("error"==e.status?e.error:"An error occurred communicating with the network.")}).catch(a)})}
+e.default=Ember.Route.extend({odnNetwork:Ember.inject.service(),model:function(e){return console.log(e),Ember.RSVP.hash({params:e,name:"sample",response:t(e.txid,this.get("odnNetwork"))}).catch(function(e){return console.error("We ran into an issue... monkeys will be dispatched to your location.",e),{msg:"Recovered from rejected promise",error:e}})},actions:{error:function(e){console.error("We ran into an issue... monkeys will be dispatched to your location.",e)}}})}),define("XXL-WebApp/services/ajax",["exports","ember-ajax/services/ajax"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})}),define("XXL-WebApp/services/odn-network",["exports","XXL-WebApp/config/environment"],function(e,t){Object.defineProperty(e,"__esModule",{value:!0})
+var n=t.default.APP.ODN_API_XXL,a=t.default.APP.ODN_API_SEARCH,s={info:"node/info",mining:"node/mining",staking:"node/staking",network:"node/network",system:"system",price:"price"}
+e.default=Ember.Service.extend({applicationName:"OdnNetwork",getInfo:function(){return new Promise(function(e,t){Ember.$.getJSON(n+"/api/"+s.info).then(function(t){e(t)}).catch(function(t){e("error grabbing info")})})},getMiningInfo:function(){return new Promise(function(e,t){Ember.$.getJSON(n+"/api/"+s.mining).then(function(t){e(t)}).catch(function(t){e("error grabbing mining")})})},getStakingInfo:function(){return new Promise(function(e,t){Ember.$.getJSON(n+"/api/"+s.staking).then(function(t){e(t)}).catch(function(t){e("error grabbing staking")})})},getNetworkInfo:function(){return new Promise(function(e,t){Ember.$.getJSON(n+"/api/"+s.network).then(function(t){e(t)}).catch(function(t){e("error grabbing network")})})},getSystemInfo:function(){return new Promise(function(e,t){Ember.$.getJSON(n+"/api/"+s.system).then(function(t){t.storageLeft=t.storage.size-t.storage.used,t.memoryLeft=t.mem.total-t.mem.available,e(t)}).catch(function(t){e("error grabbing system")})})},getPriceInfo:function(){return new Promise(function(e,t){Ember.$.getJSON(n+"/api/"+s.price).then(function(t){e(t.response)}).catch(function(t){e("error grabbing price")})})},runSearch:function(e){return new Promise(function(t,n){Ember.$.getJSON(a+"/search/"+e).then(function(e){t(e)}).catch(n)})},fetchBlock:function(e){return Ember.$.getJSON(a+"/api/block/"+e)},fetchTransaction:function(e){return Ember.$.getJSON(a+"/api/transaction/"+e)},fetchAddress:function(e){return Ember.$.getJSON(a+"/api/address/"+e)}})}),define("XXL-WebApp/templates/about",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"AHkg3mQn",block:'{"symbols":[],"statements":[[6,"div"],[9,"class","app-body"],[7],[0,"\\n  "],[1,[18,"outlet"],false],[0,"\\n  "],[6,"h2"],[7],[0,"About ObsidianXXL"],[8],[0,"\\n  "],[6,"p"],[7],[0,"The ObsidianXXL is a project started by Pixxl (Vaughn) for the Obsidian Platform for the purpose of Masternode statistics and control. It is being refined to also include the ability to view the Obsidian Blockchain that a masternode is connected too, as well display specific information about the network."],[8],[0,"\\n"],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/about.hbs"}})}),define("XXL-WebApp/templates/address",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"uSIT8yb/",block:'{"symbols":["key","value"],"statements":[[6,"div"],[9,"class","app-body"],[7],[0,"\\n  "],[6,"div"],[9,"class","hide"],[7],[0,"\\n    "],[6,"p"],[7],[0,"MSG::"],[1,[20,["model","msg"]],false],[8],[0,"\\n    "],[6,"p"],[7],[0,"Error::"],[1,[20,["model","error"]],false],[8],[0,"\\n\\n    "],[6,"p"],[7],[6,"strong"],[7],[0,"Name:"],[8],[1,[20,["model","name"]],false],[8],[0,"\\n    "],[6,"td"],[7],[1,[25,"input",null,[["type","value"],["text",[19,0,["model","name"]]]]],false],[8],[0,"\\n  "],[8],[0,"\\n\\n  "],[6,"dl"],[9,"class","hide"],[7],[0,"\\n"],[4,"each",[[25,"-each-in",[[19,0,["model","response","block"]]],null]],null,{"statements":[[0,"    "],[6,"dd"],[7],[1,[19,2,[]],false],[8],[0,"\\n    "],[6,"dt"],[7],[1,[19,1,[]],false],[8],[0,"\\n"]],"parameters":[1,2]},null],[0,"  "],[8],[0,"\\n\\n  "],[2," TODO: Animate! "],[0,"\\n  "],[6,"div"],[9,"class","xx-primary-data"],[7],[0,"\\n    "],[6,"ul"],[9,"class","clearfix"],[7],[0,"\\n      "],[6,"li"],[9,"class","u-font-size-md"],[7],[0,"type://"],[6,"span"],[7],[0,"ADDRESS"],[8],[8],[0,"\\n      "],[6,"li"],[9,"class","u-font-size-md"],[7],[0,"public_key://"],[6,"span"],[9,"class","word-wrap"],[7],[1,[20,["model","response","address","address"]],false],[8],[8],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n\\n  "],[6,"div"],[9,"class","xx-card u-margin-bottom-lg"],[7],[0,"\\n    "],[6,"div"],[9,"class","xx-data-split split-first"],[7],[0,"\\n      "],[6,"dl"],[9,"class","data-display inverse span2"],[7],[0,"\\n        "],[6,"dt"],[9,"class","overview--date"],[7],[1,[25,"format-timestamp",[[19,0,["model","response","address","last_tx"]]],null],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Last Transaction"],[8],[0,"\\n      "],[8],[0,"\\n      "],[6,"dl"],[9,"class","data-display"],[7],[0,"\\n        "],[6,"dt"],[7],[1,[20,["model","response","address","balance"]],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Balance"],[8],[0,"\\n      "],[8],[0,"\\n      "],[6,"dl"],[9,"class","data-display"],[7],[0,"\\n        "],[6,"dt"],[7],[1,[20,["model","response","address","fees"]],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Minted"],[8],[0,"\\n      "],[8],[0,"\\n      "],[6,"dl"],[9,"class","data-display"],[7],[0,"\\n        "],[6,"dt"],[7],[1,[20,["model","response","vins","length"]],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Transactions In"],[8],[0,"\\n      "],[8],[0,"\\n      "],[6,"dl"],[9,"class","data-display"],[7],[0,"\\n        "],[6,"dt"],[7],[1,[20,["model","response","vouts","length"]],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Transactions Out"],[8],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n\\n  "],[6,"h3"],[7],[0,"Transaction List coming soon..."],[8],[0,"\\n\\n\\n  "],[6,"div"],[9,"class","hide"],[7],[0,"\\n    "],[6,"h4"],[9,"class","odn-explorer--tx-list--title u-font-weight-2"],[7],[0,"Received"],[8],[0,"\\n    "],[6,"div"],[9,"class","row odn-explorer--tx-list--details clearfix"],[7],[0,"\\n      "],[6,"div"],[9,"class","col-xs-12 tx-list--details tx-list--details--vout"],[7],[0,"\\n        "],[6,"div"],[9,"class","details--vout clearfix"],[7],[0,"\\n          "],[6,"p"],[9,"class","details--address"],[7],[0,"\\n            "],[4,"link-to",["address",[19,0,["vout","address_key"]]],null,{"statements":[[1,[20,["vout","address_key"]],false]],"parameters":[]},null],[0,"\\n          "],[8],[0,"\\n          "],[6,"p"],[9,"class","details--value"],[7],[1,[20,["vout","value"]],false],[8],[0,"\\n        "],[8],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n\\n    "],[6,"h4"],[9,"class","odn-explorer--tx-list--title u-font-weight-2"],[7],[0,"Sent"],[8],[0,"\\n    "],[6,"div"],[9,"class","row odn-explorer--tx-list--details clearfix"],[7],[0,"\\n      "],[6,"div"],[9,"class","col-xs-12 tx-list--details tx-list--details--vin"],[7],[0,"\\n        "],[6,"div"],[9,"class","details--vin clearfix"],[7],[0,"\\n          "],[6,"p"],[9,"class","details--address"],[7],[0,"\\n"],[4,"if",[[25,"eq",[[19,0,["vin","address_key"]],null],null]],null,{"statements":[[0,"            Network\\n"]],"parameters":[]},{"statements":[[0,"            "],[4,"link-to",["address",[19,0,["vin","address_key"]]],null,{"statements":[[1,[20,["vin","address_key"]],false]],"parameters":[]},null],[0,"\\n"]],"parameters":[]}],[0,"          "],[8],[0,"\\n          "],[6,"p"],[9,"class","details--value pull-right"],[7],[1,[20,["vin","value"]],false],[8],[0,"\\n        "],[8],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n"],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/address.hbs"}})}),define("XXL-WebApp/templates/application",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"fUEFhZXE",block:'{"symbols":[],"statements":[[6,"div"],[9,"class","nav-mobile"],[7],[0,"\\n  "],[6,"nav"],[9,"class","nav"],[7],[0,"\\n    "],[6,"ul"],[9,"class","nav-list"],[7],[0,"\\n      "],[6,"li"],[7],[4,"link-to",["index"],[["invokeAction"],[[25,"action",[[19,0,[]],"toggleMobileNav","link"],null]]],{"statements":[[0,"\\n        Home"]],"parameters":[]},null],[8],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n"],[8],[0,"\\n\\n"],[6,"div"],[9,"class","app-view"],[7],[0,"\\n  "],[1,[25,"xxl-header",null,[["search","toggleMobileNav","version","enableSearch"],["search","toggleMobileNav",[19,0,["xxlVersion"]],[19,0,["model","enableSearch"]]]]],false],[0,"\\n\\n  "],[6,"section"],[9,"class","network"],[7],[0,"\\n"],[4,"data-display",null,[["title"],["Current Block"]],{"statements":[[0,"      "],[1,[20,["model","odnDataInfo","blocks"]],false],[0,"\\n"]],"parameters":[]},null],[0,"\\n"],[4,"data-display",null,[["title"],["Last Price"]],{"statements":[[0,"      "],[1,[25,"token-value-data-list",null,[["usd","btc"],[[19,0,["model","priceInfo","price_usd"]],[19,0,["model","priceInfo","price_btc"]]]]],false],[0,"\\n"]],"parameters":[]},null],[0,"\\n"],[4,"data-display",null,[["title"],["Network Weight"]],{"statements":[[0,"      "],[1,[25,"format-staking-weight",[[19,0,["model","odnDataMining","netstakeweight"]]],null],false],[0,"\\n"]],"parameters":[]},null],[0,"  "],[8],[0,"\\n\\n  "],[6,"section"],[10,"class",[26,["app-notification ",[25,"if",[[19,0,["model","latestRelease","update"]],"is-active",""],null]]]],[7],[0,"\\n    "],[6,"p"],[7],[0,"There is an update available for your ObsidianXXL Dashboard! You\'re currently running "],[6,"span"],[9,"class","monospace"],[7],[0,"v"],[1,[18,"xxlVersion"],false],[8],[0,", latest release is "],[6,"span"],[9,"class","monospace"],[7],[0,"v"],[1,[20,["model","latestRelease","version"]],false],[8],[0,"\\n      "],[6,"br"],[7],[8],[0,"For more information please check out the "],[6,"a"],[10,"href",[26,[[20,["model","latestRelease","link"]]]]],[9,"target","_blank"],[7],[0,"Github Release"],[8],[0,"."],[8],[0,"\\n  "],[8],[0,"\\n\\n  "],[6,"main"],[9,"class","main"],[7],[0,"\\n    "],[1,[18,"outlet"],false],[0,"\\n  "],[8],[0,"\\n\\n  "],[1,[25,"xxl-footer",null,[["version","price_usd","price_btc","currentBlock","humaizedUpdatedAt"],[[19,0,["xxlVersion"]],[19,0,["model","priceInfo","price_usd"]],[19,0,["model","priceInfo","price_btc"]],[19,0,["model","odnDataInfo","blocks"]],[19,0,["model","priceInfo","humanized_last_updated"]]]]],false],[0,"\\n"],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/application.hbs"}})}),define("XXL-WebApp/templates/block",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"aX0LOuEH",block:'{"symbols":["num","transaction","key_vout","vout","key_vin","vin","key","value"],"statements":[[6,"div"],[9,"class","app-body"],[7],[0,"\\n  "],[6,"div"],[9,"class","hide"],[7],[0,"\\n    "],[6,"p"],[7],[0,"MSG::"],[1,[20,["model","msg"]],false],[8],[0,"\\n    "],[6,"p"],[7],[0,"Error::"],[1,[20,["model","error"]],false],[8],[0,"\\n\\n    "],[6,"p"],[7],[6,"strong"],[7],[0,"Name:"],[8],[1,[20,["model","name"]],false],[8],[0,"\\n    "],[6,"td"],[7],[1,[25,"input",null,[["type","value"],["text",[19,0,["model","name"]]]]],false],[8],[0,"\\n  "],[8],[0,"\\n\\n  "],[6,"dl"],[9,"class","hide"],[7],[0,"\\n"],[4,"each",[[25,"-each-in",[[19,0,["model","response","block"]]],null]],null,{"statements":[[0,"    "],[6,"dd"],[7],[1,[19,8,[]],false],[8],[0,"\\n    "],[6,"dt"],[7],[1,[19,7,[]],false],[8],[0,"\\n"]],"parameters":[7,8]},null],[0,"  "],[8],[0,"\\n\\n  "],[2," TODO: Animate! "],[0,"\\n  "],[6,"div"],[9,"class","xx-primary-data"],[7],[0,"\\n    "],[6,"ul"],[9,"class","clearfix"],[7],[0,"\\n      "],[6,"li"],[9,"class","u-font-size-md"],[7],[0,"type://"],[6,"span"],[7],[0,"BLOCK"],[8],[8],[0,"\\n      "],[6,"li"],[9,"class","u-font-size-md"],[7],[0,"height://"],[6,"span"],[7],[1,[20,["model","response","block","height"]],false],[8],[8],[0,"\\n      "],[6,"li"],[7],[0,"hash://"],[6,"span"],[9,"class","word-wrap"],[7],[1,[20,["model","response","block","hash"]],false],[8],[8],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n\\n  "],[6,"div"],[9,"class","xx-card u-margin-bottom-lg"],[7],[0,"\\n    "],[6,"div"],[9,"class","xx-data-split split-first"],[7],[0,"\\n      "],[6,"dl"],[9,"class","data-display inverse span2"],[7],[0,"\\n        "],[6,"dt"],[9,"class","overview--date"],[7],[1,[25,"format-timestamp",[[19,0,["model","response","block","time"]]],null],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Date"],[8],[0,"\\n      "],[8],[0,"\\n      "],[6,"dl"],[9,"class","data-display"],[7],[0,"\\n        "],[6,"dt"],[7],[1,[20,["model","response","block","difficulty"]],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Difficulty"],[8],[0,"\\n      "],[8],[0,"\\n      "],[6,"dl"],[9,"class","data-display"],[7],[0,"\\n        "],[6,"dt"],[7],[1,[25,"format-bytes",[[19,0,["model","response","block","size"]]],[["unitWrap"],["true"]]],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Size"],[8],[0,"\\n      "],[8],[0,"\\n      "],[6,"dl"],[9,"class","data-display"],[7],[0,"\\n        "],[6,"dt"],[7],[1,[20,["model","response","block","mint"]],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Mint"],[8],[0,"\\n      "],[8],[0,"\\n      "],[6,"dl"],[9,"class","data-display"],[7],[0,"\\n        "],[6,"dt"],[7],[1,[20,["model","response","transactions","length"]],false],[8],[0,"\\n"],[4,"if",[[25,"or",[[25,"gt",[[19,0,["model","response","transactions","length"]],1],null],[25,"eq",[[19,0,["model","response","transactions","length"]],0],null]],null]],null,{"statements":[[0,"          "],[6,"dd"],[7],[1,[18,"title"],false],[0,"Transactions"],[8],[0,"\\n"]],"parameters":[]},{"statements":[[0,"          "],[6,"dd"],[7],[1,[18,"title"],false],[0,"Transaction"],[8],[0,"\\n"]],"parameters":[]}],[0,"      "],[8],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n\\n"],[4,"if",[[25,"not-eq",[[19,0,["model","response","block","nextblockhash"]],""],null]],null,{"statements":[[0,"  "],[6,"button"],[9,"class","hide"],[9,"type","button"],[7],[0,"Next Block"],[8],[0,"\\n"]],"parameters":[]},null],[0,"\\n"],[4,"if",[[25,"not-eq",[[19,0,["model","response","block","prevblockhash"]],""],null]],null,{"statements":[[0,"  "],[6,"button"],[9,"class","hide"],[9,"type","button"],[7],[0,"Previous Block"],[8],[0,"\\n"]],"parameters":[]},null],[0,"\\n"],[4,"each",[[25,"-each-in",[[19,0,["model","response","transactions"]]],null]],null,{"statements":[[0,"    "],[6,"div"],[9,"class","xx-card theme--action xx-card--split-grid"],[7],[0,"\\n      "],[6,"h4"],[9,"class","title u-font-weight-2"],[7],[0,"Transaction:// "],[6,"span"],[9,"class","theme-action"],[7],[1,[25,"increment-index",[[19,1,[]]],null],false],[8],[8],[0,"\\n\\n      "],[6,"div"],[9,"class","flags"],[7],[0,"\\n        "],[6,"span"],[9,"class","flags--tx-type"],[7],[1,[19,2,["tx_type"]],false],[8],[0,"\\n        "],[6,"span"],[9,"class","flags--block-type"],[7],[1,[20,["model","response","block","flags"]],false],[8],[0,"\\n      "],[8],[0,"\\n\\n      "],[6,"div"],[9,"class","details details--vin"],[7],[0,"\\n        "],[6,"h5"],[9,"class","title"],[7],[0,"Sender"],[8],[0,"\\n"],[4,"each",[[25,"-each-in",[[19,2,["vins"]]],null]],null,{"statements":[[0,"        "],[6,"div"],[9,"class","wrap clearfix"],[7],[0,"\\n          "],[6,"p"],[9,"class","details--address"],[7],[0,"\\n"],[4,"if",[[25,"eq",[[19,6,["address_key"]],null],null]],null,{"statements":[[0,"            Network\\n"]],"parameters":[]},{"statements":[[0,"            "],[4,"link-to",["address",[19,6,["address_key"]]],[["class"],["theme-block"]],{"statements":[[1,[19,6,["address_key"]],false]],"parameters":[]},null],[0,"\\n"]],"parameters":[]}],[0,"          "],[8],[0,"\\n          "],[6,"p"],[9,"class","details--value"],[7],[1,[19,6,["value"]],false],[8],[0,"\\n        "],[8],[0,"\\n"]],"parameters":[5,6]},null],[0,"      "],[8],[0,"\\n\\n      "],[6,"div"],[9,"class","details details--vout"],[7],[0,"\\n        "],[6,"h5"],[9,"class","title"],[7],[0,"Receiver"],[8],[0,"\\n"],[4,"each",[[25,"-each-in",[[19,2,["vouts"]]],null]],null,{"statements":[[4,"if",[[25,"not-eq",[[19,4,["address_key"]],null],null]],null,{"statements":[[0,"          "],[6,"div"],[9,"class","wrap clearfix"],[7],[0,"\\n            "],[6,"p"],[9,"class","details--address"],[7],[0,"\\n              "],[4,"link-to",["address",[19,4,["address_key"]]],[["class"],["theme-block"]],{"statements":[[1,[19,4,["address_key"]],false]],"parameters":[]},null],[0,"\\n            "],[8],[0,"\\n            "],[6,"p"],[9,"class","details--value"],[7],[1,[19,4,["value"]],false],[8],[0,"\\n          "],[8],[0,"\\n"]],"parameters":[]},null]],"parameters":[3,4]},null],[0,"      "],[8],[0,"\\n\\n      "],[6,"div"],[9,"class","hash"],[7],[0,"\\n        "],[6,"h5"],[9,"class","title u-font-size-md"],[7],[0,"Transaction Hash"],[8],[0,"\\n        "],[4,"link-to",["transaction",[19,2,["txid"]]],[["class"],["block-theme-highlight"]],{"statements":[[1,[19,2,["txid"]],false]],"parameters":[]},null],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n"]],"parameters":[1,2]},null],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/block.hbs"}})}),define("XXL-WebApp/templates/components/data-display",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"VkQd9qTF",block:'{"symbols":["&default"],"statements":[[6,"dt"],[7],[11,1],[8],[0,"\\n"],[6,"dd"],[7],[1,[18,"title"],false],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/components/data-display.hbs"}})}),define("XXL-WebApp/templates/components/nav-panel",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"zGDR2bRk",block:'{"symbols":["&default"],"statements":[[6,"div"],[9,"class","col-xs-12 col-sm-6"],[7],[0,"\\n  "],[6,"h4"],[7],[0,"Nav Panel"],[8],[0,"\\n"],[8],[0,"\\n\\n"],[11,1],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/components/nav-panel.hbs"}})}),define("XXL-WebApp/templates/components/odn-network",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"JKvUZpDh",block:'{"symbols":[],"statements":[[6,"div"],[9,"class","odn-network"],[7],[0,"\\n  "],[6,"div"],[9,"class","container"],[7],[0,"\\n    "],[6,"div"],[9,"class","row"],[7],[0,"\\n\\n      "],[6,"div"],[9,"class","data-display-wrap col-xs-12 col-sm-4"],[7],[0,"\\n        "],[6,"dl"],[9,"class","data-display"],[7],[0,"\\n          "],[6,"dt"],[7],[1,[20,["blockchain","blocks"]],false],[8],[0,"\\n          "],[6,"dd"],[7],[0,"Current Block"],[8],[0,"\\n        "],[8],[0,"\\n      "],[8],[0,"\\n\\n      "],[6,"div"],[9,"class","data-display-wrap col-xs-12 col-sm-4"],[7],[0,"\\n        "],[6,"dl"],[9,"class","data-display"],[7],[0,"\\n          "],[6,"dt"],[7],[1,[25,"format-staking",[[19,0,["blockchain","netstakeweight"]]],null],true],[8],[0,"\\n          "],[6,"dd"],[7],[0,"Network Weight"],[8],[0,"\\n        "],[8],[0,"\\n      "],[8],[0,"\\n\\n      "],[6,"div"],[9,"class","data-display-wrap col-xs-12 col-sm-4"],[7],[0,"\\n        "],[6,"dl"],[9,"class","data-display"],[7],[0,"\\n          "],[6,"dt"],[7],[0,"Online"],[8],[0,"\\n          "],[6,"dd"],[7],[0,"Status"],[8],[0,"\\n        "],[8],[0,"\\n      "],[8],[0,"\\n\\n    "],[8],[0,"\\n  "],[8],[0,"\\n"],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/components/odn-network.hbs"}})}),define("XXL-WebApp/templates/components/odn-node",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"kgUefsbO",block:'{"symbols":[],"statements":[[6,"h3"],[9,"class","title"],[7],[0,"Node Status"],[8],[0,"\\n"],[6,"h4"],[10,"class",[26,[[25,"if",[[19,0,["staking","staking"]],"theme-action","theme-warning"],null]]]],[7],[0,"\\n  "],[1,[25,"if",[[19,0,["staking","staking"]],"STAKING","NOT STAKING"],null],false],[0,"\\n  "],[6,"small"],[7],[0,"v1.0.0.7"],[8],[0,"\\n"],[8],[0,"\\n\\n"],[6,"div"],[9,"class","node-card node-card--light"],[7],[0,"\\n  "],[6,"div"],[9,"class","row node-card--split"],[7],[0,"\\n    "],[6,"div"],[9,"class","col-xs-12 col-sm-6"],[7],[0,"\\n      "],[6,"dl"],[9,"class","data-display inverse"],[7],[0,"\\n        "],[6,"dt"],[7],[1,[20,["info","connections"]],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Connected Peers"],[8],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n    "],[6,"div"],[9,"class","col-xs-12 col-sm-6"],[7],[0,"\\n      "],[6,"dl"],[9,"class","data-display inverse"],[7],[0,"\\n        "],[6,"dt"],[7],[1,[25,"format-staking",[[19,0,["staking","weight"]]],null],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Staking Weight"],[8],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n\\n  "],[6,"div"],[9,"class","row node-card--split"],[7],[0,"\\n    "],[6,"div"],[9,"class","col-xs-12 col-sm-6"],[7],[0,"\\n      "],[6,"dl"],[9,"class","data-display inverse"],[7],[0,"\\n        "],[6,"dt"],[7],[1,[25,"format-bytes",[[19,0,["network","totalbytessent"]]],[["unitWrap"],["true"]]],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Bytes Sent"],[8],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n    "],[6,"div"],[9,"class","col-xs-12 col-sm-6"],[7],[0,"\\n      "],[6,"dl"],[9,"class","data-display inverse"],[7],[0,"\\n        "],[6,"dt"],[7],[1,[25,"format-bytes",[[19,0,["network","totalbytesrecv"]]],[["unitWrap"],["true"]]],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Bytes Received"],[8],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n\\n  "],[6,"div"],[9,"class","row"],[7],[0,"\\n    "],[6,"div"],[9,"class","col-xs-12"],[7],[0,"\\n"],[4,"if",[[19,0,["staking","staking"]]],null,{"statements":[[0,"        "],[6,"dl"],[9,"class","data-display inverse"],[7],[0,"\\n          "],[6,"dt"],[7],[1,[25,"format-unix",[[19,0,["info","unlocked_until"]]],null],false],[8],[0,"\\n          "],[6,"dd"],[7],[0,"Staking Until"],[8],[0,"\\n        "],[8],[0,"\\n"]],"parameters":[]},{"statements":[[4,"if",[[19,0,["isReading"]]],null,{"statements":[[0,"        "],[6,"dl"],[9,"class","data-display inverse"],[7],[0,"\\n          "],[6,"dt"],[7],[0,"---"],[8],[0,"\\n          "],[6,"dd"],[7],[0,"Staking Until"],[8],[0,"\\n        "],[8],[0,"\\n      "]],"parameters":[]},null]],"parameters":[]}],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n"],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/components/odn-node.hbs"}})}),define("XXL-WebApp/templates/components/odn-server",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"vsL1FpH7",block:'{"symbols":[],"statements":[[6,"h3"],[9,"class","title"],[7],[0,"Server Status"],[8],[0,"\\n"],[6,"h4"],[9,"class","theme-action"],[7],[0,"\\n  "],[1,[20,["system","os","distro"]],false],[0," // "],[6,"span"],[7],[1,[20,["system","os","release"]],false],[8],[0,"\\n"],[8],[0,"\\n\\n"],[6,"div"],[9,"class","node-card node-card--light"],[7],[0,"\\n  "],[6,"div"],[9,"class","row node-card--split"],[7],[0,"\\n    "],[6,"div"],[9,"class","col-xs-12 col-sm-6"],[7],[0,"\\n      "],[6,"dl"],[9,"class","data-display inverse"],[7],[0,"\\n        "],[6,"dt"],[7],[0,"\\n          "],[1,[25,"format-bytes",[[19,0,["system","storageLeft"]]],[["unitWrap"],["true"]]],false],[0,"\\n        "],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Free Space"],[8],[0,"\\n      "],[8],[0,"\\n      "],[6,"dl"],[9,"class","hide data-display inverse"],[7],[0,"\\n        "],[6,"dt"],[7],[0,"\\n          "],[1,[25,"format-bytes",[[19,0,["system","storageLeft"]]],[["unitWrap"],["true"]]],false],[0," /\\n          "],[6,"span"],[9,"class","theme-action"],[7],[1,[25,"format-bytes",[[19,0,["system","storage","size"]]],[["unitWrap"],["true"]]],false],[8],[0,"\\n        "],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Free Space"],[8],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n    "],[6,"div"],[9,"class","col-xs-12 col-sm-6"],[7],[0,"\\n      "],[6,"dl"],[9,"class","data-display inverse"],[7],[0,"\\n        "],[6,"dt"],[7],[0,"\\n          "],[1,[25,"format-bytes",[[19,0,["system","memoryLeft"]]],[["unitWrap"],["true"]]],false],[0,"\\n        "],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Free Memory"],[8],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n\\n"],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/components/odn-server.hbs"}})}),define("XXL-WebApp/templates/components/side-bar",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"93vN6uTA",block:'{"symbols":[],"statements":[[6,"ul"],[9,"class","nav nav-sidebar"],[7],[0,"\\n  "],[6,"li"],[9,"class","active"],[7],[6,"a"],[9,"href","#"],[7],[0,"Overview "],[6,"span"],[9,"class","sr-only"],[7],[0,"(current)"],[8],[8],[8],[0,"\\n  "],[6,"li"],[7],[6,"a"],[9,"href","#"],[7],[0,"Reports"],[8],[8],[0,"\\n  "],[6,"li"],[7],[6,"a"],[9,"href","#"],[7],[0,"Analytics"],[8],[8],[0,"\\n  "],[6,"li"],[7],[6,"a"],[9,"href","#"],[7],[0,"Export"],[8],[8],[0,"\\n"],[8],[0,"\\n\\n"],[6,"ul"],[9,"class","nav nav-sidebar"],[7],[0,"\\n  "],[6,"li"],[7],[6,"a"],[9,"href",""],[7],[0,"Nav item"],[8],[8],[0,"\\n  "],[6,"li"],[7],[6,"a"],[9,"href",""],[7],[0,"Nav item again"],[8],[8],[0,"\\n  "],[6,"li"],[7],[6,"a"],[9,"href",""],[7],[0,"One more nav"],[8],[8],[0,"\\n  "],[6,"li"],[7],[6,"a"],[9,"href",""],[7],[0,"Another nav item"],[8],[8],[0,"\\n  "],[6,"li"],[7],[6,"a"],[9,"href",""],[7],[0,"More navigation"],[8],[8],[0,"\\n"],[8],[0,"\\n\\n"],[6,"ul"],[9,"class","nav nav-sidebar"],[7],[0,"\\n  "],[6,"li"],[7],[6,"a"],[9,"href",""],[7],[0,"Nav item again"],[8],[8],[0,"\\n  "],[6,"li"],[7],[6,"a"],[9,"href",""],[7],[0,"One more nav"],[8],[8],[0,"\\n  "],[6,"li"],[7],[6,"a"],[9,"href",""],[7],[0,"Another nav item"],[8],[8],[0,"\\n"],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/components/side-bar.hbs"}})}),define("XXL-WebApp/templates/components/token-value-data-list",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"3xtNBYuu",block:'{"symbols":["&default"],"statements":[[6,"span"],[9,"class","shrink muted"],[7],[0,"$"],[8],[0," "],[1,[25,"format-currency",[[19,0,["usd"]]],[["decimals"],["4"]]],false],[0," "],[6,"span"],[9,"class","shrink muted"],[7],[1,[18,"btc"],false],[0," Ƀ"],[8],[11,1],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/components/token-value-data-list.hbs"}})}),define("XXL-WebApp/templates/components/xxl-footer",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"4gzgg/lx",block:'{"symbols":["&default"],"statements":[[6,"div"],[9,"class","about"],[7],[0,"\\n  "],[6,"p"],[9,"class","logo"],[7],[0,"ObsidianXXL "],[6,"span"],[9,"class","version muted"],[7],[0," © 2017 | v"],[1,[18,"version"],false],[8],[8],[0,"\\n  "],[6,"ul"],[9,"class","horizontal list-social-icons"],[7],[0,"\\n    "],[6,"li"],[7],[6,"a"],[9,"target","_blank"],[9,"href","https://twitter.com/ObsidianCrypto"],[7],[6,"span"],[9,"class","icon-social-twitter icons"],[7],[8],[8],[8],[0,"\\n    "],[6,"li"],[7],[6,"a"],[9,"target","_blank"],[9,"href","https://www.facebook.com/TheObsidianPlatform/"],[7],[6,"span"],[9,"class","icon-social-facebook icons"],[7],[8],[8],[8],[0,"\\n    "],[6,"li"],[7],[6,"a"],[9,"target","_blank"],[9,"href","https://www.reddit.com/r/ObsidianProject/"],[7],[6,"span"],[9,"class","icon-social-reddit icons"],[7],[8],[8],[8],[0,"\\n  "],[8],[0,"\\n\\n  "],[6,"p"],[9,"class","u-text-align-justify--sm"],[7],[0,"ObsidianXXL is a dashboard developed for Obsidian to provide current Obsidian Node details and information about the blockchain."],[8],[0,"\\n  "],[6,"p"],[9,"class","muted"],[7],[0,"Last Price: "],[1,[25,"token-value-data-list",null,[["usd","btc"],[[19,0,["price_usd"]],[19,0,["price_btc"]]]]],false],[8],[0,"\\n  "],[6,"p"],[9,"class","muted"],[7],[0,"Price Updated: "],[1,[18,"humaizedUpdatedAt"],false],[8],[0,"\\n  "],[6,"p"],[9,"class","muted"],[7],[0,"Current Block: "],[1,[18,"currentBlock"],false],[8],[0,"\\n"],[8],[0,"\\n"],[6,"div"],[9,"class","links"],[7],[0,"\\n  "],[6,"h5"],[9,"class","title"],[7],[0,"Github Repos"],[8],[0,"\\n  "],[6,"p"],[7],[0,"ObsidianXXL "],[6,"a"],[9,"class","external"],[9,"href","https://github.com/Manbearpixel/ObsidianXXL"],[9,"target","_blank"],[7],[0,"Github"],[8],[8],[0,"\\n  "],[6,"p"],[7],[0,"Obsidian-QT Wallet "],[6,"a"],[9,"class","external"],[9,"href","https://github.com/obsidianproject/Obsidian-Qt"],[9,"target","_blank"],[7],[0,"Github"],[8],[8],[0,"\\n  "],[6,"h5"],[9,"class","title u-margin-top-sm"],[7],[0,"Obsidian Resources"],[8],[0,"\\n  "],[6,"p"],[7],[6,"a"],[9,"class","external"],[9,"href","https://obsidianplatform.com"],[9,"target","_blank"],[7],[0,"Obsidian Home"],[8],[8],[0,"\\n  "],[6,"p"],[7],[6,"a"],[9,"class","external"],[9,"href","https://medium.com/@obsidian_crypto"],[9,"target","_blank"],[7],[0,"Obsidian Blog"],[8],[8],[0,"\\n  "],[6,"p"],[7],[6,"a"],[9,"class","external"],[9,"href","https://coinmarketcap.com/currencies/obsidian/"],[9,"target","_blank"],[7],[0,"Obsidian CMC"],[8],[8],[0,"\\n"],[8],[0,"\\n"],[11,1],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/components/xxl-footer.hbs"}})}),define("XXL-WebApp/templates/components/xxl-header",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"wylkcNHD",block:'{"symbols":[],"statements":[[6,"div"],[9,"class","header--row"],[7],[0,"\\n  "],[6,"div"],[9,"class","header--logo"],[7],[0,"\\n    "],[6,"h1"],[7],[0,"ObsidianXXL"],[8],[0,"\\n    "],[6,"p"],[9,"class","version muted"],[7],[0,"v"],[1,[18,"version"],false],[8],[0,"\\n    "],[6,"div"],[9,"class","mobile-nav-action"],[3,"action",[[19,0,[]],"toggleMobileNav","action"]],[7],[0,"\\n      "],[6,"span"],[7],[8],[0,"\\n      "],[6,"span"],[7],[8],[0,"\\n      "],[6,"span"],[7],[8],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n\\n  "],[6,"nav"],[9,"class","nav"],[7],[0,"\\n    "],[6,"ul"],[9,"class","nav-list horizontal"],[7],[0,"\\n      "],[6,"li"],[7],[4,"link-to",["index"],null,{"statements":[[0,"Home"]],"parameters":[]},null],[8],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n\\n"],[4,"if",[[19,0,["enableSearch"]]],null,{"statements":[[0,"    "],[6,"div"],[9,"class","header--search"],[7],[0,"\\n      "],[6,"form"],[9,"class",""],[3,"action",[[19,0,[]],"search",[19,0,["query"]]],[["on"],["submit"]]],[7],[0,"\\n        "],[6,"div"],[10,"class",[26,["search-wrap ",[25,"if",[[19,0,["isSearchedFocused"]],"search-focus"],null]]]],[7],[0,"\\n          "],[1,[25,"input",null,[["value","class"],[[19,0,["query"]],"input text"]]],false],[0,"\\n          "],[6,"button"],[9,"type","submit"],[9,"class","button search-icon"],[7],[6,"i"],[9,"class","icon-magnifier"],[7],[8],[8],[0,"\\n        "],[8],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n"]],"parameters":[]},null],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/components/xxl-header.hbs"}})}),define("XXL-WebApp/templates/error",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"PL9mb0C+",block:'{"symbols":[],"statements":[[6,"div"],[9,"class","app-body"],[7],[0,"\\n  "],[6,"h2"],[7],[0,"An error has occurred"],[8],[0,"\\n  "],[6,"p"],[7],[0,"That is unfortunate..."],[8],[0,"\\n  "],[1,[18,"outlet"],false],[0,"\\n"],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/error.hbs"}})}),define("XXL-WebApp/templates/history",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"L6VOmGtu",block:'{"symbols":[],"statements":[[6,"div"],[9,"class","app-body"],[7],[0,"\\n  "],[1,[18,"outlet"],false],[0,"\\n  "],[6,"h2"],[7],[0,"Obsidian History"],[8],[0,"\\n  "],[6,"p"],[7],[0,"Coming soon..."],[8],[0,"\\n"],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/history.hbs"}})}),define("XXL-WebApp/templates/index",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"PXLyClQC",
+block:'{"symbols":[],"statements":[[6,"div"],[9,"class","app-body"],[7],[0,"\\n  "],[6,"h2"],[9,"class","lead"],[7],[0,"Obsidian MasterNode Dashboard"],[8],[0,"\\n\\n  "],[6,"div"],[9,"class","xx-row"],[7],[0,"\\n    "],[6,"div"],[9,"class","xx-span-6"],[7],[0,"\\n      "],[6,"h3"],[9,"class","title"],[7],[0,"Server Status"],[8],[0,"\\n      "],[6,"h4"],[9,"class",""],[7],[0,"\\n        "],[6,"span"],[9,"class","theme-action"],[7],[1,[20,["model","systemInfo","os","distro"]],false],[8],[0," "],[6,"span"],[9,"class","muted small"],[7],[1,[20,["model","systemInfo","os","release"]],false],[8],[0,"\\n      "],[8],[0,"\\n      "],[6,"div"],[9,"class","xx-card"],[7],[0,"\\n        "],[6,"div"],[9,"class","xx-data-split split-right"],[7],[0,"\\n          "],[6,"dl"],[9,"class","data-display inverse"],[7],[0,"\\n            "],[6,"dt"],[7],[0,"\\n              "],[1,[25,"format-bytes",[[19,0,["model","systemInfo","storageLeft"]]],[["unitWrap"],["true"]]],false],[0,"\\n            "],[8],[0,"\\n            "],[6,"dd"],[7],[0,"Free Space"],[8],[0,"\\n          "],[8],[0,"\\n          "],[6,"dl"],[9,"class","hide data-display inverse"],[7],[0,"\\n            "],[6,"dt"],[7],[0,"\\n              "],[1,[25,"format-bytes",[[19,0,["model","systemInfo","storageLeft"]]],[["unitWrap"],["true"]]],false],[0," /\\n              "],[6,"span"],[9,"class","theme-action"],[7],[1,[25,"format-bytes",[[19,0,["model","systemInfo","storage","size"]]],[["unitWrap"],["true"]]],false],[8],[0,"\\n            "],[8],[0,"\\n            "],[6,"dd"],[7],[0,"Free Space"],[8],[0,"\\n          "],[8],[0,"\\n          "],[6,"dl"],[9,"class","data-display inverse"],[7],[0,"\\n            "],[6,"dt"],[7],[0,"\\n              "],[1,[25,"format-bytes",[[19,0,["model","systemInfo","memoryLeft"]]],[["unitWrap"],["true"]]],false],[0,"\\n            "],[8],[0,"\\n            "],[6,"dd"],[7],[0,"Free Memory"],[8],[0,"\\n          "],[8],[0,"\\n        "],[8],[0,"\\n      "],[8],[0,"\\n\\n      "],[6,"h3"],[9,"class","title u-margin-top-sm "],[7],[0,"Network Status"],[8],[0,"\\n      "],[6,"h4"],[9,"class",""],[7],[6,"span"],[9,"class","theme-action"],[7],[1,[20,["model","odnDataInfo","connections"]],false],[8],[0," "],[6,"span"],[9,"class","muted small"],[7],[0,"Connected Peers"],[8],[0,"\\n      "],[8],[0,"\\n      "],[6,"div"],[9,"class","xx-card"],[7],[0,"\\n        "],[6,"div"],[9,"class","xx-data-split split-right"],[7],[0,"\\n          "],[6,"dl"],[9,"class","data-display inverse"],[7],[0,"\\n            "],[6,"dt"],[7],[1,[25,"format-bytes",[[19,0,["model","odnDataNetwork","totalbytessent"]]],[["unitWrap"],["true"]]],false],[8],[0,"\\n            "],[6,"dd"],[7],[0,"Bytes Sent"],[8],[0,"\\n          "],[8],[0,"\\n          "],[6,"dl"],[9,"class","data-display inverse"],[7],[0,"\\n            "],[6,"dt"],[7],[1,[25,"format-bytes",[[19,0,["model","odnDataNetwork","totalbytesrecv"]]],[["unitWrap"],["true"]]],false],[8],[0,"\\n            "],[6,"dd"],[7],[0,"Bytes Received"],[8],[0,"\\n          "],[8],[0,"\\n        "],[8],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n\\n    "],[6,"div"],[9,"class","xx-span-6"],[7],[0,"\\n      "],[6,"h3"],[9,"class","title"],[7],[0,"Node Status"],[8],[0,"\\n      "],[6,"h4"],[10,"class",[26,[[25,"if",[[19,0,["model","odnDataStaking","staking"]],"theme-action","theme-warning"],null]]]],[7],[0,"\\n        "],[1,[25,"if",[[19,0,["model","odnDataStaking","staking"]],"STAKING","NOT STAKING"],null],false],[0,"\\n        "],[6,"small"],[7],[1,[25,"format-node-version",[[19,0,["model","odnDataInfo","version"]]],null],false],[8],[0,"\\n      "],[8],[0,"\\n      "],[6,"div"],[9,"class","xx-card"],[7],[0,"\\n        "],[6,"div"],[10,"class",[26,["xx-data-split split-right ",[25,"if",[[19,0,["model","odnDataStaking","staking"]],"split-first",""],null]]]],[7],[0,"\\n"],[4,"if",[[19,0,["model","odnDataStaking","staking"]]],null,{"statements":[[0,"            "],[6,"dl"],[9,"class","data-display span2"],[7],[0,"\\n              "],[6,"dt"],[9,"class","u-text-transform-capitalize"],[7],[1,[25,"format-staking-time",[[19,0,["model","odnDataStaking","expectedtime"]]],[["unitWrap"],["true"]]],false],[8],[0,"\\n              "],[6,"dd"],[7],[0,"Estimated Reward"],[8],[0,"\\n            "],[8],[0,"\\n\\n            "],[6,"dl"],[9,"class","data-display span2"],[7],[0,"\\n              "],[6,"dt"],[7],[1,[25,"format-unix",[[19,0,["model","odnDataInfo","unlocked_until"]]],null],false],[8],[0,"\\n              "],[6,"dd"],[7],[0,"Staking Until"],[8],[0,"\\n            "],[8],[0,"\\n            "],[6,"hr"],[9,"class","split-bottom"],[7],[8],[0,"\\n"]],"parameters":[]},null],[0,"          "],[6,"dl"],[9,"class","data-display inverse"],[7],[0,"\\n            "],[6,"dt"],[7],[1,[25,"format-staking-weight",[[19,0,["model","odnDataStaking","weight"]]],null],false],[8],[0,"\\n            "],[6,"dd"],[7],[0,"Staking Weight"],[8],[0,"\\n          "],[8],[0,"\\n          "],[6,"dl"],[9,"class","data-display inverse"],[7],[0,"\\n            "],[6,"dt"],[7],[1,[20,["model","odnDataInfo","balance"]],false],[8],[0,"\\n            "],[6,"dd"],[7],[0,"Balance"],[8],[0,"\\n          "],[8],[0,"\\n        "],[8],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n\\n"],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/index.hbs"}})}),define("XXL-WebApp/templates/search",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"SEdrDXiO",block:'{"symbols":[],"statements":[[6,"div"],[9,"class","app-body"],[7],[0,"\\n  "],[6,"div"],[9,"class","container"],[7],[0,"\\n    "],[6,"div"],[9,"class","col-xs-12"],[7],[0,"\\n      "],[6,"h2"],[7],[0,"Obsidian Search"],[8],[0,"\\n      "],[6,"h4"],[7],[0,"Search for "],[6,"span"],[9,"class","theme-action"],[7],[1,[20,["model","params","query"]],false],[8],[0," returned empty."],[8],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n"],[8],[0,"\\n"],[1,[18,"outlet"],false],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/search.hbs"}})}),define("XXL-WebApp/templates/stats",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"5+5OpU8R",block:'{"symbols":[],"statements":[[6,"div"],[9,"class","app-body"],[7],[0,"\\n  "],[1,[18,"outlet"],false],[0,"\\n  "],[6,"h2"],[7],[0,"Obsidian Network Stats"],[8],[0,"\\n  "],[6,"p"],[7],[0,"Coming soon..."],[8],[0,"\\n"],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/stats.hbs"}})}),define("XXL-WebApp/templates/transaction",["exports"],function(e){Object.defineProperty(e,"__esModule",{value:!0}),e.default=Ember.HTMLBars.template({id:"z1FFXApl",block:'{"symbols":["key_vout","vout","key_vin","vin","key","value"],"statements":[[6,"div"],[9,"class","app-body"],[7],[0,"\\n  "],[6,"div"],[9,"class","hide"],[7],[0,"\\n    "],[6,"p"],[7],[0,"MSG::"],[1,[20,["model","msg"]],false],[8],[0,"\\n    "],[6,"p"],[7],[0,"Error::"],[1,[20,["model","error"]],false],[8],[0,"\\n\\n    "],[6,"p"],[7],[6,"strong"],[7],[0,"Name:"],[8],[1,[20,["model","name"]],false],[8],[0,"\\n    "],[6,"td"],[7],[1,[25,"input",null,[["type","value"],["text",[19,0,["model","name"]]]]],false],[8],[0,"\\n  "],[8],[0,"\\n\\n  "],[6,"dl"],[9,"class","hide"],[7],[0,"\\n"],[4,"each",[[25,"-each-in",[[19,0,["model","response","block"]]],null]],null,{"statements":[[0,"    "],[6,"dd"],[7],[1,[19,6,[]],false],[8],[0,"\\n    "],[6,"dt"],[7],[1,[19,5,[]],false],[8],[0,"\\n"]],"parameters":[5,6]},null],[0,"  "],[8],[0,"\\n\\n  "],[2," TODO: Animate! "],[0,"\\n  "],[6,"div"],[9,"class","xx-primary-data"],[7],[0,"\\n    "],[6,"ul"],[9,"class","clearfix"],[7],[0,"\\n      "],[6,"li"],[9,"class","u-font-size-md"],[7],[0,"type://"],[6,"span"],[7],[0,"TRANSACTION"],[8],[8],[0,"\\n      "],[6,"li"],[9,"class","u-font-size-md"],[7],[0,"tx_type://"],[6,"span"],[7],[1,[20,["model","response","transaction","tx_type"]],false],[8],[8],[0,"\\n      "],[6,"li"],[7],[0,"txid://"],[6,"span"],[9,"class","word-wrap"],[7],[1,[20,["model","response","transaction","txid"]],false],[8],[8],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n\\n  "],[6,"div"],[9,"class","xx-card u-margin-bottom-lg"],[7],[0,"\\n    "],[6,"div"],[9,"class","xx-data-split split-first"],[7],[0,"\\n      "],[6,"dl"],[9,"class","data-display inverse span2"],[7],[0,"\\n        "],[6,"dt"],[9,"class","overview--date"],[7],[1,[25,"format-timestamp",[[19,0,["model","response","transaction","time"]]],null],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Date"],[8],[0,"\\n      "],[8],[0,"\\n      "],[6,"dl"],[9,"class","data-display"],[7],[0,"\\n        "],[6,"dt"],[7],[1,[20,["model","response","transaction","height_key"]],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Height"],[8],[0,"\\n      "],[8],[0,"\\n      "],[6,"dl"],[9,"class","data-display"],[7],[0,"\\n        "],[6,"dt"],[7],[1,[20,["model","response","fee"]],false],[8],[0,"\\n        "],[6,"dd"],[7],[0,"Fees"],[8],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n  "],[8],[0,"\\n\\n\\n  "],[6,"div"],[9,"class","xx-card theme--action"],[7],[0,"\\n    "],[6,"h4"],[9,"class","title u-font-weight-2"],[7],[0,"Sender"],[8],[0,"\\n\\n"],[4,"each",[[25,"-each-in",[[19,0,["model","response","vins"]]],null]],null,{"statements":[[0,"    "],[6,"div"],[9,"class","details details--vin"],[7],[0,"\\n      "],[6,"div"],[9,"class","wrap clearfix"],[7],[0,"\\n        "],[6,"p"],[9,"class","details--address"],[7],[0,"\\n"],[4,"if",[[25,"eq",[[19,4,["address_key"]],null],null]],null,{"statements":[[0,"          Network\\n"]],"parameters":[]},{"statements":[[0,"          "],[4,"link-to",["address",[19,4,["address_key"]]],[["class"],["theme-block"]],{"statements":[[1,[19,4,["address_key"]],false]],"parameters":[]},null],[0,"\\n"]],"parameters":[]}],[0,"        "],[8],[0,"\\n        "],[6,"p"],[9,"class","details--value"],[7],[1,[19,4,["value"]],false],[8],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n"]],"parameters":[3,4]},null],[0,"  "],[8],[0,"\\n\\n  "],[6,"div"],[9,"class","xx-card theme--action"],[7],[0,"\\n    "],[6,"h4"],[9,"class","title u-font-weight-2"],[7],[0,"Receiver"],[8],[0,"\\n\\n"],[4,"each",[[25,"-each-in",[[19,0,["model","response","vouts"]]],null]],null,{"statements":[[0,"    "],[6,"div"],[9,"class","details details--vout"],[7],[0,"\\n      "],[6,"div"],[9,"class","wrap clearfix"],[7],[0,"\\n        "],[6,"p"],[9,"class","details--address"],[7],[0,"\\n"],[4,"if",[[25,"eq",[[19,2,["address_key"]],null],null]],null,{"statements":[[0,"          Network\\n"]],"parameters":[]},{"statements":[[0,"          "],[4,"link-to",["address",[19,2,["address_key"]]],[["class"],["theme-block"]],{"statements":[[1,[19,2,["address_key"]],false]],"parameters":[]},null],[0,"\\n"]],"parameters":[]}],[0,"        "],[8],[0,"\\n        "],[6,"p"],[9,"class","details--value"],[7],[1,[19,2,["value"]],false],[8],[0,"\\n      "],[8],[0,"\\n    "],[8],[0,"\\n"]],"parameters":[1,2]},null],[0,"  "],[8],[0,"\\n"],[8],[0,"\\n"]],"hasEval":false}',meta:{moduleName:"XXL-WebApp/templates/transaction.hbs"}})}),define("XXL-WebApp/config/environment",["ember"],function(e){try{var t="XXL-WebApp/config/environment",n=document.querySelector('meta[name="'+t+'"]').getAttribute("content"),a=JSON.parse(unescape(n)),s={default:a}
+return Object.defineProperty(s,"__esModule",{value:!0}),s}catch(e){throw new Error('Could not read config from meta tag with name "'+t+'".')}}),runningTests||require("XXL-WebApp/app").default.create({enable_search:!1,ODN_API_XXL:"",ODN_API_SEARCH:"http://localhost:3000",name:"XXL-WebApp",version:"0.1.1+"})
