+++ public/index.html
@@ -15,7 +15,7 @@
     <link href="https://fonts.googleapis.com/css?family=Source+Code+Pro:300,400" rel="stylesheet">
 
     <link rel="stylesheet" href="/assets/vendor-f1649674d31f9a962b7e691628d00185.css">
-    <link rel="stylesheet" href="/assets/XXL-WebApp-de4e001a51788d106bc6ce81b9507630.css">
+    <link rel="stylesheet" href="/assets/XXL-WebApp-8311fc3e43e6ccdb12a3c7ceb4bbb03c.css">
 
     
   </head>
@@ -26,6 +26,6 @@
     
 
     <script src="/assets/vendor-b1b2e6cc4c411593ca6bb9ea2d768e68.js"></script>
-    <script src="/assets/XXL-WebApp-9922518d5f48dbc8e6407c02e76ab6e9.js"></script>
+    <script src="/assets/XXL-WebApp-d09ec568f7e96b796d488cf060517436.js"></script>
   </body>
 </html>
